server/requirements.txt
pycrypto==2.6.1
git+https://github.com/Edorka/blockpy
